chore: Update APKINFO configuration files
This commit updates the configuration files related to APKINFO. It renames the `cli-inputs.json` file in the `auth` directory, changing the identity pool and user pool names to match the new APKINFO version. It also removes the `amplifyhosting-template.json` file from the `hosting` directory. Additionally, it updates the `index.html` file to change the page title to "APK-INFO".